Move livewire tournament components into own namespace
