+++ resources/views/livewire/tournament/register.blade.php
